Fix: Add calendar userHook
