+++ src/components/navigation/navibarComponents/homeNavigation.tsx
@@ -11,6 +11,7 @@ import { ko } from 'date-fns/locale'
 import CalendarMenu from '@/components/navigation/navibarButtons/calendarMenu'
 import TravelDesButton from '@/components/navigation/navibarButtons/travelDesButton'
 import OneLayerNavibar from '@/components/navigation/navibarComponents/oneLayerNavibar'
+import useCalendarLogic from '@/components/navigation/navibarButtons/calendarUseHook'
 
 export type PersonType = 'adult' | 'child' | 'baby' | 'pet'
 
@@ -22,11 +23,13 @@ export interface Person {
 }
 // named type ->
 export default function HomeNavigation() {
+  const ref = useRef<any>(null)
   const buttonsizeboolen = true
+
   const [searchButtonHover, setSearchButtonHover] = useState(false)
   const [activeButton, setActiveButton] = useState(0)
   const [topActivityMenu, setTopActivityMenu] = useState(true)
-  const ref = useRef<any>(null)
+
   const [inputValue, setInputValue] = useState('')
 
   const [isMenuOpen, setIsMenuOpen] = useState(false)
@@ -39,6 +42,7 @@ export default function HomeNavigation() {
     baby: 0,
     pet: 0,
   })
+
   const personSetter = (type: PersonType) => {
     return {
       plus: () => {
@@ -58,20 +62,17 @@ export default function HomeNavigation() {
   }
 
   //달력 로직
-  const [calendarOpen, setCalendarOpen] = useState(false)
-  const defaultSelected: DateRange = {
-    from: undefined,
-    to: undefined,
-  }
-  const [range, setRange] = useState<DateRange | undefined>(defaultSelected)
-  const [plusDate, setPlusDate] = useState('')
-  const [plusdateClick, setPlusDateClick] = useState(0)
-
-  const handleCalendar = () => {
-    setRange(defaultSelected)
-    setPlusDate('')
-    setPlusDateClick(0)
-  }
+  const {
+    calendarOpen,
+    setCalendarOpen,
+    range,
+    setRange,
+    plusDate,
+    setPlusDate,
+    plusDateClick,
+    setPlusDateClick,
+    handleCalendar,
+  } = useCalendarLogic()
 
   const handleNumber = () => {
     const resetNumber = { adult: 0, child: 0, baby: 0, pet: 0 }
@@ -275,7 +276,7 @@ export default function HomeNavigation() {
               setCalendarOpen={setCalendarOpen}
               setPlusDate={setPlusDate}
               setPlusDateClick={setPlusDateClick}
-              plusdateClick={plusdateClick}
+              plusdateClick={plusDateClick}
             />
           </div>
           {/* 날짜 선택 끝 */}
